Fix relative imports causing infinite recursion.
If two files in separate directories import each other using relative paths then the registered path for each file grows and grows.

To resolve this we determine and register the absolute uri for each loaded schema and use this to check for already processed files.

Add Uri handling class from Zend Framework as a dependency to help resolving relative urls and include this in generated phar file. Other Zend Framework dependencies are expected to be added later on when moving to a third party code generation library.

Implemented unit test.